+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## [NEXT]
+## [0.6.18]
 * BREAKING CHANGES: change `forceTileSize` type double to Size.
 * Add support to offsetX and offsetY in TileMap layers.